Move utils to own file & handle non-multipart forms
