+++ frontend/src/pages/RequestorPage/RequestPanel/CopyAsCurl/index.ts
@@ -1,0 +1,1 @@
+export * from "./CopyAsCurl";
